+++ src/core/stats.cc
@@ -2,8 +2,8 @@
 
 namespace rosewood { namespace core { namespace stats {
 
-    Counter<int> draw_calls;
-    Counter<int> triangle_count;
-    Counter<int> shader_change_count;
+    Counter<size_t> draw_calls;
+    Counter<size_t> triangle_count;
+    Counter<size_t> shader_change_count;
 
 } } }
